Tiny update
Added two new computers in office. (Cactus guy starting to have a objective)

Added interger array to notepads to differentiate them. This makes it easier to add objectives to indvidual
notepads.

Small objective (Not finished was mainly used for testing)

Few changes made to email widget in an attempt to make the email buttons text be the same as the headings text (unuccessful so far)